+++ CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 
 # Unreleased
 
-- [`Fix`] fixed `trasnferFungibleAsset` function 
+- [`Fix`] fixed `trasnferFungibleAsset` function
 - Run all examples in CI
+- Introcude cli `Move` class that holds `move` related commands
+- Add common cli commands - `move.init()`, `move.compile()`, `move.test()`, `move.publish()`
 
 # 1.13.3 (2024-04-30)